+++ App.js
@@ -5,6 +5,7 @@ import Weather from './components/weather';
 import { FontAwesome5 } from '@expo/vector-icons'; 
 import * as Location from 'expo-location';
 import Forecast from './components/Forecast';
+import WeatherContainer from './components/WeatherContainer';
 
 const API_URL = 'http://192.168.100.19:5000/weather';
 
@@ -23,155 +24,45 @@ const defaultState = {
 }
 
 
+
 export default function App() {
   
-  const [state, setState] = useState(defaultState)
-  const [searchCity, setSearchCity] = useState('')
+
   const [errorMessage, setErrorMessage] = useState('')
-  const [currentLocation, setCurrentLocation] = useState(null)
+  const [coordinates,setCoordinates] = useState(null)
 
-  
-
-
-
-
- 
-
-  const fetchData = async () => {
-
+  const getPermisions = async ()=> {
+    let {status} = await Location.requestPermissionsAsync();
+    if(status !== 'granted'){
+      setErrorMessage('Permision to access location was denied');
+      return
+    }
+    let location = await Location.getCurrentPositionAsync({});
+    setCoordinates({lat: location.coords.latitude,lon: location.coords.longitude})
+    console.log(coordinates)
     
-    try{
-     
-  
-      const data = await fetch(
-        `${API_URL}?location=${state.location}`
-      )
-   
-      if(data.status===200){
-        const jsonData = await data.json()
-      
-      setState({
-        ...jsonData,
-        location: jsonData.location? jsonData.location : state.location,
-        ready: true
-   
-      })
-      setErrorMessage('')
-
-      }else {
-        const message = await data.json()
-        setState({
-          ...state,
-          ready: false,
-        })
-        setErrorMessage(message.message)
-      }
 
 
-      
-     
-  
-
-    } catch (error){
-      console.log(error);
-      setErrorMessage('error')
-
-    } 
-    
-    
-    
   }
 
-  const submit = () => {
-    console.log(searchCity);
-    setState( {
-      
-        ...state,
+  useEffect(()=>{
+    getPermisions()
+  },[])
+
+
+  if(coordinates){
    
-        location: searchCity
-      
-    })
+   
+    return (
+      <WeatherContainer lat={coordinates.lat} lon={coordinates.lon}/>   
+    );
   }
-
-
- 
-  
-  useEffect(()=> {    
-  
-      if(state.location){
-        fetchData()
-      }
-   
-      
-  }, [state.location]) 
-
-
-
   return (
+    <Text>
+      {errorMessage}
+    </Text>
+  )
 
-    <View style={styles.mainContainer}>
-      <View style={styles.searchContainer}>
-          <TextInput
-            style={styles.input} 
-            placeholder="city"
-            
-            onChangeText={input => setSearchCity(input)}
-            onSubmitEditing={(e)=>submit()}
-            value={searchCity}
-            />  
-          <TouchableOpacity onPress={submit}>
-              <FontAwesome5 name="search" size={24} color="black" />
-          </TouchableOpacity> 
-
-      </View>
-
-      <View style={styles.container}>    
-        {errorMessage? <Text>{errorMessage}</Text> : state.ready ?  <Weather  {...state} /> : <View></View>}
-        {state.ready && <Forecast lon={state.lon} lat={state.lat} />}        
-      </View> 
-     
-    </View>
-   
-  );
+ 
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 6,
-   
-    alignItems: 'center',
-    
-    
-    
-  },
-  searchContainer: {
-    flex: 1,
-    flexDirection: 'row',
-
-    justifyContent: 'center',
-    marginTop: 20
-  
-
-
-  },
-  mainContainer: {
-    marginTop: '8%',
-    flex: 1,
-    flexDirection: 'column',
-    alignContent: 'center',
-    justifyContent: 'space-evenly',
-    marginBottom: '4%',
-  },
-  input: {
-    width: 90,
-    borderColor: 'gray',
-    borderWidth: 2,
-    height: 35,
-    borderRadius: 5,
-    marginRight: 10,
-    marginTop: 1,
-    
-  },
-
-
-});
